Reduce integration test execution time
