+++ include-build/roborazzi-gradle-plugin/src/integrationTest/java/io/github/takahirom/roborazzi/RoborazziGradleProject.kt
@@ -385,5 +385,55 @@ class RoborazziTest {
     )
   }
 
+  fun addTestCaptureWithCustomPathTest() {
+    val file =
+      testProjectDir.root.resolve("app/src/test/java/com/github/takahirom/integration_test_project/RoborazziTest.kt")
+    file.parentFile.mkdirs()
+    file.writeText(
+      """
+package com.github.takahirom.integration_test_project
+
+import androidx.test.core.app.ActivityScenario
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.matcher.ViewMatchers
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.takahirom.roborazzi.*
+import android.app.Application
+import androidx.test.core.app.ApplicationProvider
+import org.robolectric.Shadows
+import android.content.ComponentName
+import org.junit.Test
+import org.junit.Rule
+import org.robolectric.annotation.GraphicsMode
+
+import org.junit.Assert.*
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+@GraphicsMode(GraphicsMode.Mode.NATIVE)
+class RoborazziTest {
+
+  @get:Rule val roborazziRule = RoborazziRule()
+  init {
+    com.github.takahirom.roborazzi.ROBORAZZI_DEBUG = true
+  }
+  
+  @Test
+  fun testCaptureWithCustomPath() {
+    ActivityScenario.launch(MainActivity::class.java)
+    onView(ViewMatchers.isRoot()).captureRoboImage(
+      filePath = "${"$"}DEFAULT_ROBORAZZI_OUTPUT_DIR_PATH/customdir/custom_file.png",
+      roborazziOptions = RoborazziOptions(
+        compareOptions = RoborazziOptions.CompareOptions(
+          outputDirectoryPath = "${"$"}DEFAULT_ROBORAZZI_OUTPUT_DIR_PATH/custom_compare_outputDirectoryPath",
+        )
+      )
+    )
+  }
+}
+    """
+    )
+  }
+
   val appBuildFile = AppBuildFile(testProjectDir)
 }
